feat: frontend and backend integration test
