+++ spx-gui/.env.staging
@@ -1,6 +1,7 @@
 # Config for env staging
 
-VITE_API_BASE_URL="https://goplus-builder.qiniu.io/api"
+# VITE_API_BASE_URL="https://goplus-builder.qiniu.io/api"
+VITE_API_BASE_URL="https://1ace-103-119-224-183.ngrok-free.app"
 VITE_VERCEL_PROXIED_API_BASE_URL="https://goplus-builder.qiniu.io/api"
 
 VITE_CASDOOR_ENDPOINT="https://goplus-casdoor.qiniu.io"
